feat(translateProvider): add a preferredLanguage property
Add the "preferredLanguage" property in order to solve an initial language
setting issue. At the configuration phase it's better to use this property
instead of the "uses" method, because if "rememberLanguage" is set to true
it may invoke a language file loading process twice (the language from the
uses method and a stored language).