feat(storageKey): add a storageKey method
The storageKey method allows the user to change a key, which will be used
to store the language code. So, the user can change this key during the
configuration phase now.

BREAKING CHANGE: The $STORAGE_KEY isn't represent a current storage key
from now. To discover which key is used now you have to call the storageKey
method without params.