feat(translateProvider): add methods to use different missingTranslationHandlers
This commit adds  useMissingTranslationHandler and  useMissingTranslationHandlerLog (as a shortcut for the former).

It also introduces a new factory service  which could then be used in application using  useMissingTranslateHandlerLog().

It is now possible to add any kind of missing translation handler and tell angular-translate to use it.

BREAKING CHANGES:  missingTranslationHandler is no longer supported since its functionality will be replaced with  useMissingTranslationHandlerLog.