test(invalidate): fix tests for the invalidate() method
To make it possibility to reconfigure module in different "describe" areas we have to
inject $translate service directly when want to use it. In other case we cannot change
module configuration due injector is already called (or something like that).

Custom loader was rewritten to simulate multiple languages loading properly.