feat(getLangInfo): Adds an ability store info about the language in the translation table
A method getLangInfo() of the $translate service allows the user to retrieve some
predefined data about the language from it's translation table.

The predefined data is stored in the special predefined namespace in the translations
table. User is able to change the name of the predefined namespace by using
a langInfoNamespace() method of the provider.