feat(): add an option for post processing compiling
Using the optional directive attribute `translate-compile`, the result of a translation will be recompiled.

For example:

```json
{
  'text': '<span>{{name}} is a citizen of <strong ng-bind="world"></strong>!</span>'
}
```

```html
<p translate="text" translate-compile translate-values="{name: \'The Doctor\'}"></p>
```

this will result into (given a scope variable `world = 'Gallifrey'` the element text `The Doctor is a citizen of Gallifrey!` respectively in the corresponding HTML.

Starting now, this feature is enabled globally. This feature can be disabled via `$translateProvider.usePostCompiling(false)`; it can be read with `$translate.isPostCompilingEnabled()`. You can also enable or disable it locally per directive attribute.

1. if `$translate.isPostCompilingEnabled() === false`: use `translate-compile` within a directive to enable it locally.
2. if `$translate.isPostCompilingEnabled() === true`: use `translate-compile="false"` within a directive to disable it locally.