feat (loaders): enable loaders cache
It useful if we disable cache in our project (for dynamic content), but want enable it for translation files (static content).
Also it useful if we need include translation into build file (cache in advance for production).

app.run(function($translationCache) {
    $translationCache.put('/compiled/module1/locals/en.json', '{"MODULE1.HELLO":"Hi!"}');
    // Other translations string added by Grunt
});
In conclusion, we never need uncached translations, because it is fixed content.