Adding ability to register a transformer to manipulate the static file received and transform them into JSON object.
The $translateStaticFilesLoader has been transformed into a provider so we can now
register a specific transformer in the case where the static files loaded from the
server would not be valid JSON object. For instance, if the files are .properties
(used for Java i18n) the file could be transformed into a valid key-value pair
object which would keep the compatibility with angular-translate.