feat($translateStaticFilesLoader): add a transformer for static files
The $translateStaticFilesLoader has been transformed into a provider so we can now
register a specific transformer to keep using angular-translate even if the files are
not valid JSON objects. When people come from a Java background and this could be used
to offer a way to transform the .properties files into a valid JSON object.

// How to use it :
angular.module(...).config(function($translateStaticFilesLoaderProvider) {
  $translateStaticFilesLoaderProvider.setTransformer(function(receivedData) {
    // Handle the data
  });
});