fix(test/refresh): fix current table refreshing test
The old version of the test is not appropriate due to logical errors. Test was build to work with
en_EN language while de_DE was specified, so no changes were actually done to the current
translation table. The test is extended to check that old translation keys are deleted and new
translation keys are added by the refresh method.