feat(directive): Add translation IDs to the scope
This change allows to use translation IDs extracted by `translate`
directive in other directives.

For example, if you are developing a translation editor, you can create
a directive, let’s call it `angular-translate-edit`, with `restrict:
‘A’`, that would get translation IDs defined for the angular-translate.

This gives you the possibility to avoid some complex code duplication
with all different options to define translation IDs, including
interpolation, what is already done in the `translate` directive.

As a part of this change one duplicated test is removed that uses
`$rootScope.translationIds` due to the name conflict in the
prototypically inherited directive scope. The same test already exists
and it just uses `$rootScope.translationId` instead.