fix(service): fix interpolation issue with non-string as input
Since AngularJS 1.5, the input values (i.e. of a filter) will used 'as is' which means numbers (as translation ids) will not be transformed to strings anymore. This generates unexpected errors.

Because this fix, the result is that explicit typed number values will be return (transformed as a string).

```
// before and < AJS 1.5
$filter('translate)(0) === 0
// after and < AJS 1.5
$filter('translate)(0) === '0'

// before and AJS 1.5+
$filter('translate)(0) exception
// after and AJS 1.5+
$filter('translate)(0) === '0'

```

Fixes #1511