fix($translate): reassign language promises in refresh
When using $translatePartialLoader, the language promise is invalid upon adding a new part and calling $translate.refresh; calling $translate immediately after adding a new part results in the language key being returned from the already-resolved promise that was created on the original invokation of loadAsync, rather than waiting for the new part to be fetched from the server. Calling $translate.use with $forceAsyncReloadEnabled=true configured is a workaround with the current code, but it is logically incorrect to force a language change when a refresh is the appropriate action. This change remedies the problem by assigning the promise from loadAsync to the appropriate langPromise upon refresh.