fix($translate): update translation tables at the appropriate time in refresh and simplify the routine
With my previous fix in-place, it was revealed that that the asynchronous processing queue (aka $q, promise) generated by aggregating the results of the individual language promises in the branch where all languages are refreshed executes the logic to update the translation tables after each of the individual table promises is already resolved and their queues have been processed. This means that, when using the partial-loader, any calls to $translate for strings contained in a newly-added part that are made during the refresh procedure will erroneously have a cache miss because the determineTranslation call is waiting on the individual table promises, which resolve all of their handlers (promise.then) before the outer handler ($q.all().then) begins processing its handler to actually update the translation tables. This is resolved by moving the translation table update handler to the individual event queues for each table, thereby guaranteeing the table has been updated by the time determineTranslation is executed because the handlers are executed in the order they are registered and the table-updating handler will be the first thing registered. After really digging in and understanding refresh I also noticed that there was a lot of duplicate and overly-complex logic, so I refactored it to be much simpler.