feat(loader-partial): addPart specific urlTemplate override
To accommodate use case in which location of translation table for a
partial is not known at config time and is unique to partial being
requested. For example, app is served as a SPA from s3 bucket, after
initialization and based on route parameters addParts from other s3
buckets and/or paths based on those route parameters.