fix: Allow line breaks in interpolated expressions.
Line breaks can now appear in interpolated AngularJS expressions inside
the body or translated elements and inside the value of translate
attributes.

Also needed to remove `fsevents` from `package.json` for `npm install`
to work on Linux.  Might want to add that back in as an
"optionalDependancy" if it is really needed.

Closes #1884, #1824