feat: throw TypeError for translationId
AngularJS 1.6 introduced a new behaviour that caused an exception on unhandled rejections of promises. If you were to pass undefined to to $translate the error message would simply be "Possibly unhandled rejection: undefined". With this change a TypeError is thrown instead so that you can easly find where the invalid translationId was passed from in the stack trace

feat: throw TypeError for translationId

AngularJS 1.6 introduced a new behaviour that caused an exception on unhandled rejections of promises. If you were to pass undefined to to $translate the error message would simply be "Possibly unhandled rejection: undefined". With this change a TypeError is thrown instead so that you can easly find where the invalid translationId was passed from in the stack trace